booking_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic<2
aio-pika
python-dotenv